+++ CHANGELOG.md
@@ -7,10 +7,15 @@ _Maintainer | Robin Verachtert | robin.verachtert@froomle.com_
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [0.3.6]
 ### Bugfixes
 * __datasets__
     * Fixed bug in MovieLens dataset download
+
+### Additions
+* __datasets__
+    * Added Globo dataset
 
 ### Changes
 * __datasets__